+++ UnityProject/Assets/Scripts/Health/Living/SimpleAnimal/SimpleAnimal.cs
@@ -31,7 +31,7 @@ public class SimpleAnimal : HealthBehaviour
 		SetAliveState(deadState);
 	}
 
-	public override int ReceiveAndCalculateDamage(GameObject damagedBy, int damage, DamageType damageType,
+	public override float ReceiveAndCalculateDamage(GameObject damagedBy, float damage, DamageType damageType,
 		BodyPartType bodyPartAim)
 	{
 		base.ReceiveAndCalculateDamage(damagedBy, damage, damageType, bodyPartAim);
