+++ UnityProject/Assets/Scripts/Messages/Server/UpdateUIMessage.cs
@@ -8,7 +8,7 @@ using UnityEngine.Networking;
 public class UpdateUIMessage : ServerMessage
 {
 	public static short MessageType = (short) MessageTypes.UpdateUIMessage;
-	public int CurHealth;
+	public float CurHealth;
 	public NetworkInstanceId Recipient;
 
 	public override IEnumerator Process()
@@ -24,7 +24,7 @@ public class UpdateUIMessage : ServerMessage
 	/// </summary>
 	/// <param name="recipient">Recipient.</param>
 	/// <param name="cHealth">Current server health.</param>
-	public static UpdateUIMessage Send(GameObject recipient, int cHealth)
+	public static UpdateUIMessage Send(GameObject recipient, float cHealth)
 	{
 		UpdateUIMessage msg = new UpdateUIMessage
 		{
